I think the refactor/reorg is mostly done, I'm happy with it
